Refactor: Move NeRF helper utilities to kornia.geometry.camera.utils (#3365)
* Moved NeRF test helpers into kornia.testing (part of tests reorg) + run pre-commit fixes

* Move nerf helpers to kornia.geometry.camera.utils and update function signature

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add docstrings and fix pre-commit issues

* Fix type annotations in camera utils for mypy compliance

* Apply pre-commit formatting fixes

---------
