+++ testing/nerf.py
@@ -1,0 +1,26 @@
+# LICENSE HEADER MANAGED BY add-license-header
+#
+# Copyright 2018 Kornia Team
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+from __future__ import annotations
+
+# Shim imports for backward compatibility - implementations moved to kornia.geometry.camera.utils
+from kornia.geometry.camera.utils import (
+    create_pinhole_camera,
+)
+
+# Deprecated alias for backward compatibility - use create_pinhole_camera instead
+create_one_camera = create_pinhole_camera
