+++ tests/nerf/test_nerf_solver.py
@@ -20,13 +20,16 @@ import torch
 
 from kornia.core import Tensor
 from kornia.geometry.camera import PinholeCamera
+from kornia.geometry.camera.utils import (
+    create_four_cameras,
+    create_pinhole_camera,
+    create_random_images_for_cameras,
+    create_red_images_for_cameras,
+)
 from kornia.nerf.nerf_model import NerfModelRenderer
 from kornia.nerf.nerf_solver import NerfSolver
 
 from testing.base import assert_close
-
-from tests.nerf.test_data_utils import create_random_images_for_cameras, create_red_images_for_cameras
-from tests.nerf.test_rays import create_four_cameras, create_one_camera
 
 
 class TestNerfSolver:
@@ -50,7 +53,7 @@ class TestNerfSolver:
 
     def test_only_red_uniform_sampling(self, device, dtype):
         torch.manual_seed(1)  # For reproducibility of random processes
-        camera: PinholeCamera = create_one_camera(5, 9, device, dtype)
+        camera: PinholeCamera = create_pinhole_camera(5, 9, device, dtype)
         img: list[Tensor] = create_red_images_for_cameras(camera, device)
 
         # train the model
@@ -65,7 +68,7 @@ class TestNerfSolver:
         assert_close(img_rendered.to(device, dtype), img[0].to(device, dtype) / 255.0)
 
     def test_single_ray(self, device, dtype):
-        camera: PinholeCamera = create_one_camera(5, 9, device, dtype)
+        camera: PinholeCamera = create_pinhole_camera(5, 9, device, dtype)
         img: list[Tensor] = create_red_images_for_cameras(camera, device)
 
         nerf_obj = NerfSolver(device=device, dtype=dtype)
@@ -75,7 +78,7 @@ class TestNerfSolver:
     def test_only_red(self, device, dtype):
         torch.manual_seed(0)  # For reproducibility of random processes
 
-        camera: PinholeCamera = create_one_camera(5, 9, device, dtype)
+        camera: PinholeCamera = create_pinhole_camera(5, 9, device, dtype)
         img: list[Tensor] = create_red_images_for_cameras(camera, device)
 
         # train the model
